+++ acc_app_optimisation/gui/__init__.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
 """Module containing the GUI elements of this application."""
 
+from ._translate import InitialSelection
 from .main_window import MainWindow